♻️ moving HTTP client into crawler/client.rs module
